Initial commit.
> I don't always code, but when I do, I do it in prod.  Both for fun and profit.

I'm going to regret doing this, I know it.  But, here it is.

If this ended in `.sh` I think it'd get less scrutiny.  Committing something to
begin actually pushing this into shape.

History: this was cobbled together on the 4th of July during an incident and deployed
with the scripts found in scrits/.  Good bad and/or ugly, here we are.