Limit the number of concurrent WAL files that we process.
Unlike previously where we would blindly run around and fault all the
WAL files all the time and limit the number of IOs per WAL file, the new
structure is to limit the number of concurrent IOs and govern the number
of outstanding IOs but process the IOs sequentially per WAL file.

The next commit should contain the IO request throttling.