Make the DB retry logic conditional.
For process and service managers that retry indefinitely, the preferred
behavior is to fail initialization, thereby signalling a failure to the
manager.  For service managers that proctively disable a service this
may be undesirable, however.

The `run` command now supports a --retry-db-init as well as a .toml config
option, `run.retry-db-init` to enable this retry behavior.

This behavior is broken and predicated on #17 being fixed.