Rework the LSN type to be actually correct 100% of the time.
Previously there were several bogus assumptions in the math and my
understanding of the way an LSN was constructed and used.  Through
the magic of concatenating two `uint32`'s into a `uint64`, everything
"worked", but not correctly.

Huge thank you to @gwydirsam for sitting with me a few times acting
as my sanity filter and sounding board.