Enable WAL prefaulting for PostgreSQL processes that are in recovery.
If the database is running, connect to the database and pre-fault per
information contained in the database.

If the database is starting up, however, we're probably in a world of
hurt and would like the prefaulter to pre-fault in the necessary pages
in order to speed up the rate at which the database becomes available

When `pg_prefaulter` is unable to connect to the database and recognizes
the database is starting up, `pg_prefaulter` resorts to *NIX CLI
utilities to parse out the WAL file being replayed according to the
process titles.  Hardware fails.  Databases misbehave.  The need to
recovery quickly is important.

Fixes: #20