Enable WAL prefaulting for PostgreSQL processes that are in recovery. (#25)
* Remove redundant check for shutdown: this is handled in the event loop

* Spell `--retry-db-init` correctly to match the docs and config file.

* Start PostgreSQL using exec in order to reduce one layer of sh(1)

* Reorganize the `pg` package.  Spawn `pg/types.go` for everything not LSN.

* Fix the locale of the SMF manifest.

Pointed out by: @jclulow

* Pull out the WAL query into a standalone function.

* Rename pg.New() to pg.NewLSN()

When the `pg` package was named `lsn` the old name made more sense.

* Use the correct return type for pg.WALFileName().

* Rename WALFileName() to WALFilename()

* Evaluate the deferred log call in an anonymous function.

This prevents the serialization of arguments at the time the defer
is evaluated.  This prevents mysterious shutdown messages from
appearing in the past.

Noticed by: @jclulow

* Use bytes as the unit of readahead for prefaulting WAL records.

Previous config value: `wal-readahead`
New config value: `wal-readahead-bytes`

Measuring this in segments was confusing in a bunch of places.

Also, it's "readahead" and not "read-ahead" or "readAhead".

* Add an AddBytes() helper method to pg.LSN.

* Add lsn.Readahead(bytes): a helper method that returns a slice of WALFilenames

* Add pg.LSNCmp() helper function to compare two LSNs.

* Push around tests into their respective locations.

* Add pg.ParseWALFilename() to parse an LSN and TimelineID from a WAL filename

* Rename `l` to `lsn` now that the package name has changed to `pg`.

* Use Errorf() instead of Fatalf() when testing.

* Remove redundant testing that is already present in pg/types_test.go

* Improve the AddBytes test and actually compare the `equalLSN` value

* Improve the testing used in ParseWALFilename.

* Add a MustParseLSN() helper function.

* Beef up the testing for the LSNCmp() method.

* Rework the LSN type to be actually correct 100% of the time.

Previously there were several bogus assumptions in the math and my
understanding of the way an LSN was constructed and used.  Through
the magic of concatenating two `uint32`'s into a `uint64`, everything
"worked", but not correctly.

Huge thank you to @gwydirsam for sitting with me a few times acting
as my sanity filter and sounding board.

* Enable WAL prefaulting for PostgreSQL processes that are in recovery.

If the database is running, connect to the database and pre-fault per
information contained in the database.

If the database is starting up, however, we're probably in a world of
hurt and would like the prefaulter to pre-fault in the necessary pages
in order to speed up the rate at which the database becomes available

When `pg_prefaulter` is unable to connect to the database and recognizes
the database is starting up, `pg_prefaulter` resorts to *NIX CLI
utilities to parse out the WAL file being replayed according to the
process titles.  Hardware fails.  Databases misbehave.  The need to
recovery quickly is important.

Fixes: #20

* Commit miss: Remove entries from the cache when there is an error.