Add a guard against a close on an already closed handle.
This didn't result in a panic, but it did result in a mismatch in the
purge FD invariant where the close counter could be one higher than
the open count.  There is a small window of time where an error can
occur and PostgreSQL goes away and initiates a shutdown wherein this
condition could occur.