Illumos performance improvements (#29)
* Use `/proc` on Illumos to detect the proc args

* Ignore database OID 0

Database OID 0 is used when creating a database and is a magic OID
that the prefaulter should ignore.

* Database shutdown is now a recoverable error.

* Move the purge invariant to a more reasonable place *cough* correct *cough*

* Fix pargs(1).  New log output:

```
{"time":"2017-09-30T20:33:48.302207541Z","level":"debug","walfile":"000000010000000000000060","message":"found WAL segment from pargs(1)"}
{"time":"2017-09-30T20:33:49.690287117Z","level":"debug","walfile":"000000010000000000000061","message":"found WAL segment from pargs(1)"}
```

* Fix `/proc` PID parsing

* Remove `ps(1)` support from Illumos.

* Retry if PostgreSQL is starting up and extracting the pid from `/proc` fails

* Add a guard against a close on an already closed handle.

This didn't result in a panic, but it did result in a mismatch in the
purge FD invariant where the close counter could be one higher than
the open count.  There is a small window of time where an error can
occur and PostgreSQL goes away and initiates a shutdown wherein this
condition could occur.

* Bump the FD ttl down to 5min.

In the even that we haven't accessed a file in 5min, I'm okay `close(2)`'ing
the FD and taking the hit to re-`open(2)` it.

* Teach `pg_prefaulter(1)` how to read from `pg_xlogdump(1)` using `-f`

This should reduce the amount of fork/exec that gets called on a
caught-up database.

* Use gauge metrics vs histograms in a few places.

It turns out the histogram metrics for these values isn't that interesting
because there isn't that much variance.