Refactor QueryOldestLSN into QueryOldestLSNs.
There are several LSNs that may be significant.  When dealing with
apply lag, the `pg_last_xlog_replay_location()` is probably the most
important LSN in your vocabulary.  However, the `redo_location` is
also important in that if the pages referenced by the `redo_location`
are not part of your filesystem cache.

If the pages referenced by the WAL file pointed to by the `redo_location`
are not in the filesystem cache, the `CHECKPOINT` process will potentially
fall behind permanently and the number of WAL files present in `pg_xlog/`
will become unbounded.

See also joyent/pg_prefaulter#30 for a discussion of prefaulting the WAL
file used by `pg_last_xlog_receive_location()` in order to improve the
performance of `synchronous_commit=remote_write`.