De-dupe WAL file processing activity.
Previously I thought this was happening in the LRU cache, but I was
mistaken.  Now, before faulting in a WAL file we instead check to
make sure the WAL file is not already in progress.  If it is, we
skip processing the file.  If the page is already present, similarly,
we assume it's "InProgress" because from the caller's perspective,
the result is indistinguishable.