Use a sync.WaitGroup as opposed to cmd.Wait()
The former is correct, whereas the latter presumes there wouldn't be
a problem with the scanner.  This is substantially more rubust.

Random self-trivia that I won't forget any time soon: the `Add()`
call to the `sync.WaitGroup` must happen before the spawn of the
`go`.  I naively treated it like a Lock and put the `Add()`/`Done()`
inside of the `go` routine, but the `Wait()` call didn't have
anything to wait for, so it happily returned immediately.
Very silly, totally obvious, and subtle unless you're looking for it.